+++ src/main/java/Application.java
@@ -1,10 +1,9 @@
-import controller.BlackjackConsole;
 import view.InputView;
 import view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
-        BlackjackConsole blackjackController = new BlackjackConsole(new OutputView(), new InputView());
-        blackjackController.run();
+        BlackjackConsole blackjackConsole = new BlackjackConsole(new OutputView(), new InputView());
+        blackjackConsole.run();
     }
 }
