+++ src/main/java/card/CardDeck.java
@@ -1,4 +1,4 @@
-package domain;
+package card;
 
 import java.util.ArrayList;
 import java.util.List;
