+++ src/main/java/card/CardRank.java
@@ -1,4 +1,4 @@
-package domain;
+package card;
 
 public enum CardRank {
     ACE(1, "A"),
