+++ src/main/java/game/GameResult.java
@@ -1,4 +1,4 @@
-package domain;
+package game;
 
 public enum GameResult {
     WIN,
