+++ src/test/java/domain/BettingTableTest.java
@@ -1,9 +1,9 @@
 package domain;
 
-import domain.user.Dealer;
-import domain.user.Participants;
-import domain.user.Player;
-import domain.user.User;
+import user.Dealer;
+import user.Participants;
+import user.Player;
+import user.User;
 import java.util.List;
 import java.util.Map;
 import org.assertj.core.api.Assertions;
