+++ src/test/java/domain/user/PlayerTest.java
@@ -1,13 +1,14 @@
 package domain.user;
 
-import domain.Card;
-import domain.CardDeck;
+import card.Card;
+import card.CardDeck;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import user.Player;
 
 class PlayerTest {
     @DisplayName("일반 유저는 자신이 가진 모든 카드를 공개해야 한다")
