+++ src/test/java/domain/user/UserTest.java
@@ -1,9 +1,11 @@
 package domain.user;
 
-import domain.CardDeck;
+import card.CardDeck;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import user.Player;
+import user.User;
 
 class UserTest {
     @DisplayName("카드를 뽑았을 때 21을 넘기면 버스트된다")
